Refactor player stats and event handling
- Updated the GetMatches, GetLeaderboard, GetPlayerStats, GetPlayerMatches, and GetPlayerPerformanceHistory methods to convert match_id to string and adjust duration to float64.
- Changed PlayerCount and Kills types in MatchSummary from int to uint64 for better compatibility with database types.
- Modified the initial schema to include round_number in the raw_events table and created a new player_stats_daily table for unified player stats.
- Removed the obsolete migration script for adding round_number.
- Added new tools for debugging ClickHouse queries and verifying tokens, including map popularity tests and server token retrieval.
- Implemented a SHA256 hashing function for token verification.

+++ internal/models/events.go
@@ -196,8 +196,8 @@ type MatchSummary struct {
 	Map         string    `json:"map"`
 	StartTime   time.Time `json:"start_time"`
 	Duration    float64   `json:"duration"`
-	PlayerCount int       `json:"player_count"`
-	Kills       int       `json:"kills"`
+	PlayerCount uint64    `json:"player_count"`
+	Kills       uint64    `json:"kills"`
 }
 
 // RawEvent is the incoming event from game servers
